GitHub copilot instructions update (#180)
## Purpose
<!-- Describe the intention of the changes being proposed. What problem
does it solve or functionality does it add? -->
* ...

## Does this introduce a breaking change?
<!-- Mark one with an "x". -->
```
[ ] Yes
[x] No
```

## Pull Request Type
What kind of change does this Pull Request introduce?

Updating the 03_openai page, based on my review of the UI. The "Apply in
editor" button previously mentioned doesn't appear for me. Clicking the
"Apply" button in the top right of the code box either opens the VS.Code
command box with the command to create the new file, or updates an
existing file. I've updated the second screenshot and associated text to
account for this.

<!-- Please check the one that applies to this PR using "x". -->
```
[ ] Bugfix
[ ] Feature
[ ] Code style update (formatting, local variables)
[ ] Refactoring (no functional changes, no api changes)
[x] Documentation content changes
[ ] Other... Please describe:
```

## How to Test


https://ronline-java-aca-preview.azurewebsites.net/docs/03_lab_openai/03_openai.html

## What to Check
Verify updates remain accurate